Randomization + Main profile for WH added
Added function for randomizing values as well as the Main Profile generation for Warhammer. This is now usable and I will test this next Sunday in practice in a gaming session. For future Shadowrun needs to be added and maybe the possibility to generate many characters into a file as well as further information for WH (secondary profile? Equipment?).